refactor: move audit logger import path to audit.logger_db
