+++ backend/app/api/routes/firebase_config.py
@@ -17,7 +17,7 @@ from app.schemas.firebase_config import (
     FirebaseTestResponse,
 )
 from app.services.notifications.firebase import FirebaseService
-from app.services.audit_logger_db import get_audit_logger_db
+from app.services.audit.logger_db import get_audit_logger_db
 from app.core.rate_limiter import user_limiter, get_limit
 
 logger = logging.getLogger(__name__)
