+++ backend/app/api/routes/logging.py
@@ -11,7 +11,7 @@ from app.api.deps import get_current_user, get_db
 from app.core.config import settings
 from app.schemas.user import UserPublic
 from app.schemas.audit_log import AuditLogResponse
-from app.services.audit_logger_db import get_audit_logger_db
+from app.services.audit.logger_db import get_audit_logger_db
 from app.services.disk_monitor import get_disk_io_history
 from app.core.rate_limiter import user_limiter, get_limit
