+++ backend/app/api/routes/rate_limit_config.py
@@ -14,7 +14,7 @@ from app.schemas.rate_limit_config import (
     RateLimitConfigList
 )
 from app.services.rate_limit_config import RateLimitConfigService
-from app.services.audit_logger_db import get_audit_logger_db
+from app.services.audit.logger_db import get_audit_logger_db
 from app.core.rate_limiter import user_limiter, get_limit
 
 router = APIRouter(prefix="/rate-limits", tags=["admin", "rate-limits"])
