+++ backend/app/api/routes/updates.py
@@ -30,7 +30,7 @@ from app.schemas.update import (
     VersionHistoryEntry,
 )
 from app.services.update.api import get_update_service, get_update_backend
-from app.services.audit_logger_db import get_audit_logger_db
+from app.services.audit.logger_db import get_audit_logger_db
 
 router = APIRouter(prefix="/updates", tags=["updates"])
